Build against available metamod and all possible sdks
With the current version of the script, it’s not easy for a beginner to compile a sample mm plugin.

In fact, we need to set when we configure the build the SDK and the mm path. It’s not very clear for beginners + the script crash with an exception which cause a confusion

With this version, just make build directory, do a configure in direction then, build => enjoy!